refactor browser launching logic and use better process forking (#252)
* refactor browser launching logic and use better process forking

shifts browser launching logic to pkg/launcher and removes a lot of the nested function calls which were making the 'assume' command code difficult to read.

* spacing

* handle forkprocess on windows

* fix lint error

* fix lint

* fix build
