+++ pkg/browser/linux.go
@@ -1,4 +1,4 @@
-package browsers
+package browser
 
 import (
 	"os/exec"
